Implement Phase 3: quasiquote, gensym, file load, boot file
Add quasiquote compiler (`, ,, ,@) as special form — handles atoms,
unquote, splicing via APPEND, dotted tails, and nested lists using
existing opcodes (no new ones). Add gensym builtin for hygienic macros.

Add platform_file_read() for both POSIX and AmigaOS, (load "path")
builtin, and boot file loading (lib/boot.lisp) at startup providing
cadr/caar/cdar/cddr/caddr/cadar/identity/endp.

Fix vbcc warnings: replace (void)x with CL_UNUSED() macro, use %lu
with (unsigned long) casts in mem.c. Amiga build is now warning-free.

68 host tests, 184 Amiga batch tests — all passing.

+++ src/core/mem.c
@@ -371,9 +371,10 @@ void cl_gc(void)
 void cl_mem_stats(void)
 {
     char buf[256];
-    sprintf(buf, "Heap: %u/%u bytes used, %u free, %u GC cycles\n",
-            cl_heap.total_allocated, cl_heap.arena_size,
-            cl_heap.arena_size - cl_heap.total_allocated,
-            cl_heap.gc_count);
+    sprintf(buf, "Heap: %lu/%lu bytes used, %lu free, %lu GC cycles\n",
+            (unsigned long)cl_heap.total_allocated,
+            (unsigned long)cl_heap.arena_size,
+            (unsigned long)(cl_heap.arena_size - cl_heap.total_allocated),
+            (unsigned long)cl_heap.gc_count);
     platform_write_string(buf);
 }
